Split rig module into submodules
